Include bench-timing.h only if _ISOMAC isn't defined
bench-timing.h includes hp-timing.h for high precision timing functions.
But hp-timing.h is an internal header file, which can't be included when
_ISOMAC is defined.  But _ISOMAC is defined for C++ benchmark programs
via libc-symbols.h.  We don't include bench-timing.h if _ISOMAC is
defined.  If high precision timing functions are needed in C++ benchmark
programs, we need to revisit this issue.

	* benchtests/bench-string.h: Include bench-timing.h only if
	_ISOMAC isn't defined.